requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
djangorestframework==3.12.1
importlib-metadata==2.0.0
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
zipp==3.4.0
